Complete Management
Database management complete. Sales screen implemented. Validation
added.

+++ EPOS/PaymentMethod.cs
@@ -32,8 +32,6 @@ namespace EPOS
         {
             Globals.GetColors();
             BackColor = Color.FromName(Globals.Backcolor);
-            listBox1.BackColor = Color.FromName(Globals.Backcolor);
-            listBox1.ForeColor = Color.FromName(Globals.Fontcolor);
             labelName.Text = Globals.Pubname;
             foreach (Control c in this.Controls)
             {
@@ -98,13 +96,27 @@ namespace EPOS
                 switch (dr)
                 {
                     case DialogResult.Yes:
-                        int id = int.Parse(listBox1.SelectedValue.ToString());
-                        SqlConnection cat = new SqlConnection(Globals.dataconnection);
-                        cat.Open();
-                        SqlCommand delete = new SqlCommand("DELETE FROM Payment WHERE PaymentID = @id", cat);
-                        delete.Parameters.AddWithValue("@id", id);
-                        delete.ExecuteNonQuery();
-                        cat.Close();
+                        if (int.Parse(listBox1.SelectedValue.ToString()) != 1)
+                        {
+                            try
+                            {
+                                int id = int.Parse(listBox1.SelectedValue.ToString());
+                                SqlConnection cat = new SqlConnection(Globals.dataconnection);
+                                cat.Open();
+                                SqlCommand delete = new SqlCommand("DELETE FROM Payment WHERE PaymentID = @id", cat);
+                                delete.Parameters.AddWithValue("@id", id);
+                                delete.ExecuteNonQuery();
+                                cat.Close();
+                            }
+                            catch
+                            {
+                                MessageBox.Show("This payment method cannot be deleted as it is currently in use");
+                            }
+                        }
+                        else
+                        {
+                            MessageBox.Show("Can not delete Cash as a payment method");
+                        }
                         break;
                     case DialogResult.No: break;
                 }
